Fix: add the allow_speaker function back
Seems like in the last commit this function was deleted but was still being referenced later in the code and that is messing up the demo. I am not entirely sure if the goal of the last commit was to completely get rid of it, but it seems that the reference wasn't deleted so I just added it back.